refactor: introduce Sink trait and break slack.rs into module
Split the 1,327-line slack.rs into a slack/ module with focused files:
- blocks.rs: Block Kit type definitions + Serialize impls
- markdown.rs: markdown-to-blocks and markdown-to-mrkdwn converters
- tests.rs: all ~30 slack tests consolidated
- mod.rs: SlackWebhookSink + SlackApiSink implementing new Sink trait

Replace CronTrigger's private ResolvedSink enum with Arc<dyn Sink>,
simplifying execute_once() to a single sink.deliver() call. Remove
the dead deliver() function from sinks/mod.rs.
