+++ src/tasks/triggers/cron.rs
@@ -9,21 +9,17 @@ use croner::Cron;
 use crate::config::{SinkConfig, SourceConfig};
 use crate::tasks::context::render_prompt;
 use crate::tasks::executors::Executor;
-use crate::tasks::sinks;
+use crate::tasks::sinks::Sink;
+use crate::tasks::sinks::slack::{SlackApiSink, SlackWebhookSink};
 use crate::tasks::sources::{self, ContentItem};
 
 pub struct CronTrigger {
     cron: Cron,
     sources: Vec<SourceConfig>,
-    sink: Option<ResolvedSink>,
+    sink: Option<Arc<dyn Sink>>,
     working_dir: PathBuf,
     http_client: Arc<reqwest::Client>,
     github_token: Option<String>,
-}
-
-enum ResolvedSink {
-    SlackWebhook { webhook_url: String },
-    SlackApi { bot_token: String, channel: String },
 }
 
 impl CronTrigger {
@@ -40,7 +36,7 @@ impl CronTrigger {
             .map_err(|e| anyhow::anyhow!("invalid cron expression '{}': {}", schedule, e))?;
 
         // Resolve sink env vars at startup so we fail fast
-        let resolved_sink = match sink {
+        let resolved_sink: Option<Arc<dyn Sink>> = match sink {
             Some(SinkConfig::Slack {
                 webhook_url_env,
                 bot_token_env,
@@ -53,12 +49,19 @@ impl CronTrigger {
                     let channel = channel.with_context(|| {
                         "slack bot_token_env requires a channel to be set"
                     })?;
-                    Some(ResolvedSink::SlackApi { bot_token, channel })
+                    Some(Arc::new(SlackApiSink::new(
+                        Arc::clone(&http_client),
+                        bot_token,
+                        channel,
+                    )))
                 } else if let Some(webhook_env) = webhook_url_env {
                     let webhook_url = std::env::var(&webhook_env).with_context(|| {
                         format!("sink requires env var {webhook_env} but it is not set")
                     })?;
-                    Some(ResolvedSink::SlackWebhook { webhook_url })
+                    Some(Arc::new(SlackWebhookSink::new(
+                        Arc::clone(&http_client),
+                        webhook_url,
+                    )))
                 } else {
                     anyhow::bail!("slack sink requires either webhook_url_env or bot_token_env");
                 }
@@ -158,23 +161,9 @@ impl CronTrigger {
         // 5. Deliver to sink if configured
         if let Some(sink) = &self.sink {
             if !exec_result.text.is_empty() {
-                match sink {
-                    ResolvedSink::SlackWebhook { webhook_url } => {
-                        sinks::slack::post_to_url(&self.http_client, webhook_url, &exec_result.text)
-                            .await
-                            .context("failed to deliver to Slack webhook")?;
-                    }
-                    ResolvedSink::SlackApi { bot_token, channel } => {
-                        sinks::slack::post_threaded_blocks(
-                            &self.http_client,
-                            bot_token,
-                            channel,
-                            &exec_result.text,
-                        )
-                        .await
-                        .context("failed to deliver to Slack API")?;
-                    }
-                }
+                sink.deliver(&exec_result.text)
+                    .await
+                    .context("failed to deliver to sink")?;
             } else {
                 tracing::warn!(task = %task_name, "Executor returned empty text, skipping sink delivery");
             }
